Add code to generate validator ed25519 address used in genesis config
